requirements.txt
cookies==2.2.1
issue==0.1.5
oauthlib==1.0.3
py==1.4.31
pytest==2.9.1
PyYAML==3.11
requests==2.9.1
requests-oauthlib==0.6.1
responses==0.5.1
six==1.10.0
tweepy==3.5.0
